# Conflicts:
#	hotel-common/src/main/java/com/gt/hotel/base/BaseController.java
#	hotel-web/src/main/java/com/gt/hotel/interceptor/MobileAuthenticationInterceptor.java

+++ hotel-entity/src/main/java/com/gt/hotel/other/jxc/Product.java
@@ -17,7 +17,7 @@ public class Product {
 
     private int buyPrice;
 
-    private int createDate;
+    private long createDate;
     
     private String createDateStr;
