+++ hotel-entity/src/main/java/com/gt/hotel/other/jxc/Unit.java
@@ -15,7 +15,7 @@ import lombok.Data;
 @Data
 public class Unit {
 
-    private int createDate;
+    private long createDate;
 
     private String createDateStr;
